Extract evaluation logic
