+++ main_WEEKEND.py
@@ -8,100 +8,8 @@ from tg_bot import send_telegram_message
 
 from models import ActorCritic
 from worker import worker
-from config import CURRENCY_CONFIGS, TradingConfig, set_global_seed
-from simulated_env import SimulatedOandaForexEnv
-
-
-def evaluate_model(
-    model,
-    currency_config,
-    episodes: int = 3,
-    candle_count: int = 5000,
-):
-    """Evaluate a model in a simulated environment using the full candle set.
-
-    This runs each episode over ``candle_count`` candles instead of a short
-    fixed number of steps. The evaluation uses greedy action selection
-    by choosing the action with the highest predicted probability. Additional
-    statistics such as profit factor and maximum drawdown are printed for a
-    more thorough assessment.
-    """
-
-    env = SimulatedOandaForexEnv(
-        currency_config,
-        candle_count=candle_count,
-        granularity=TradingConfig.GRANULARITY,
-    )
-
-    model.eval()
-    total_reward = 0.0
-    total_trades = 0
-    profits = []
-    action_counts = [0, 0, 0]  # long, short, neutral
-
-    for _ in range(episodes):
-        state = torch.tensor(env.reset(), dtype=torch.float32).unsqueeze(0)
-        decisions = torch.zeros((1, 16), dtype=torch.float32)
-        episode_reward = 0.0
-
-        done = False
-        while not done:
-            with torch.no_grad():
-                logits, _ = model(state, decisions)
-                probs = torch.softmax(logits, dim=1)
-                action = torch.argmax(probs, dim=1).item()
-
-            action_counts[action] += 1
-            next_state, reward, done, _ = env.step(action)
-            episode_reward += reward
-
-            if done or next_state is None:
-                break
-
-            state = torch.tensor(next_state, dtype=torch.float32).unsqueeze(0)
-
-        total_reward += episode_reward
-        profits.extend([t.profit for t in env.trade_log])
-        total_trades += len(env.trade_log)
-
-    avg_reward = total_reward / episodes
-    if profits:
-        avg_profit = sum(profits) / len(profits)
-        win_rate = sum(p > 0 for p in profits) / len(profits)
-        positive = sum(p for p in profits if p > 0)
-        negative = abs(sum(p for p in profits if p < 0))
-        profit_factor = positive / negative if negative else float("inf")
-        cumulative = torch.tensor(profits).cumsum(0)
-        max_cum = torch.maximum(cumulative, cumulative.clone().cummax(0)[0])
-        drawdown = (max_cum - cumulative).max().item()
-    else:
-        avg_profit = 0.0
-        win_rate = 0.0
-        profit_factor = 0.0
-        drawdown = 0.0
-
-    total_actions = sum(action_counts)
-    if total_actions > 0:
-        action_dist = [c / total_actions for c in action_counts]
-    else:
-        action_dist = [0.0, 0.0, 0.0]
-
-    print(f"Evaluation result for {currency_config.instrument}:")
-    print(f"  Avg reward per episode: {avg_reward:.4f}")
-    print(
-        f"  Trades: {total_trades}, Win rate: {win_rate*100:.1f}%, "
-        f"Avg profit: {avg_profit:.4f}"
-    )
-    print(
-        "  Action distribution: "
-        f"long {action_dist[0]*100:.1f}%, "
-        f"short {action_dist[1]*100:.1f}%, "
-        f"neutral {action_dist[2]*100:.1f}%"
-    )
-    print(f"  Profit factor: {profit_factor:.2f}")
-    print(f"  Max drawdown: {drawdown:.4f}")
-
-    return avg_reward, avg_profit, win_rate
+from config import CURRENCY_CONFIGS, set_global_seed
+from evaluation import evaluate_model
